requirements.txt
django
django-allauth==0.29.0
django-crispy-forms
django-imagekit
django-notifications-hq
django-simple-captcha
django-pagedown
django-activity-stream
django-bootstrap-pagination

mistune
Pillow
bleach